backend/requirements.dev.txt
-r requirements.txt
django-debug-toolbar==4.3.0
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
flake8==7.0.0
django-debug-toolbar==4.3.0